procedural_human/pyproject.toml
[project]
name = "procedural-human"
version = "1.0.0"
description = "Procedural Human Generator for Blender - Create realistic human models using Geometry Nodes"
readme = "README.md"
requires-python = "==3.11.*"
authors = [
    { name = "Procedural Human Team", email = "info@proceduralhuman.dev" },
]
keywords = [
    "blender",
    "3d",
    "procedural",
    "human",
    "geometry-nodes",
    "modeling",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: Multimedia :: Graphics :: 3D Modeling",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
    "fake-bpy-module>=20251003",
    "tree-sitter>=0.20.0",
    "tree-sitter-python>=0.20.0",
    "filemover @ git+https://github.com/plex1/filemover",
    "torch @ https://download.pytorch.org/whl/cu118/torch-2.5.1%2Bcu118-cp311-cp311-win_amd64.whl ; sys_platform == 'win32'",
    "torchvision @ https://download.pytorch.org/whl/cu118/torchvision-0.20.1%2Bcu118-cp311-cp311-win_amd64.whl ; sys_platform == 'win32'",
    "transformers @ git+https://github.com/huggingface/transformers.git",
    "timm>=1.0.17",
    "tqdm",
    "ftfy==6.1.1",
    "regex",
    "iopath>=0.1.9",
    "huggingface_hub",
    "imageio",
    "opencv-python==4.11.0.86",
    "open3d",
    "fastapi",
    "uvicorn",
    "requests",
    "typer",
    "pillow",
    "antlr4-python3-runtime==4.9.2",
    "omegaconf<2.4",
    "evo",
    "e3nn",
    "moviepy==1.0.3",
    "plyfile",
    "pillow_heif",
    "safetensors",
    "addict",
    "pycolmap",
    "ultralytics",
    "lapx>=0.5.5",
    "pandas",
    "prettytable",
    "numba",
    "pypose",
    "scikit-learn",
    "scikit-fmm",
    "depth-anything-3 @ git+https://github.com/ByteDance-Seed/Depth-Anything-3.git",
]

[project.optional-dependencies]
cuda = []

[tool.uv.extra-build-dependencies]
      xformers = ["torch"]

[project.urls]
Homepage = "https://github.com/procedural-human/blender-addon"
Documentation = "https://procedural-human.readthedocs.io/"
Repository = "https://github.com/procedural-human/blender-addon.git"
"Bug Tracker" = "https://github.com/procedural-human/blender-addon/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.hatch.metadata]
allow-direct-references= true

[tool.hatch.build.targets.wheel]
      packages = ["."]